# server/docs/swagger.yaml
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: 400
          type: integer
        error:
          example: validation failed
          type: string
        message:
          example: Bad request
          type: string
      type: object
...
    dto.AuthResponseDTO:
      properties:
        bootstrap_admin:
          type: boolean
        expiresAt:
          type: string
        mfa_methods:
          items:
            type: string
          type: array
          uniqueItems: false
        mfa_token:
          type: string
        refreshToken:
          type: string
        requires_mfa:
          type: boolean
        token:
          type: string
        user:
          $ref: '#/components/schemas/dto.UserDTO'
      type: object
...
    dto.DisableTOTPRequestDTO:
      properties:
        current_password:
          type: string
      required:
      - current_password
      type: object
...
    dto.EnableTOTPRequestDTO:
      properties:
        code:
          type: string
        setup_token:
          type: string
      required:
      - code
      - setup_token
      type: object
...
    dto.MFAStatusDTO:
      properties:
        available_methods:
          items:
            type: string
          type: array
          uniqueItems: false
        passkey_count:
          type: integer
        recovery_codes_generated_at:
          type: string
        recovery_codes_remaining:
          type: integer
        totp_enabled:
          type: boolean
      type: object
...
    dto.RecoveryCodesResponseDTO:
      properties:
        generated_at:
          type: string
        recovery_codes:
          items:
            type: string
          type: array
          uniqueItems: false
        status:
          $ref: '#/components/schemas/dto.MFAStatusDTO'
      type: object
...
    dto.RegenerateRecoveryCodesRequestDTO:
      properties:
        current_password:
          type: string
      required:
      - current_password
      type: object
...
    dto.TOTPSetupResponseDTO:
      properties:
        account_name:
          type: string
        issuer:
          type: string
        otpauth_uri:
          type: string
        secret:
          type: string
        setup_token:
          type: string
      type: object
...
    dto.UserDTO:
      properties:
        avatar_asset_id:
          type: string
        created_at:
          type: string
        display_name:
          type: string
        is_active:
          type: boolean
        last_login:
          type: string
        permissions:
          items:
            type: string
          type: array
          uniqueItems: false
        role:
          type: string
        updated_at:
          type: string
        user_id:
          type: integer
        username:
          type: string
      type: object
...
    dto.VerifyMFARequestDTO:
      properties:
        code:
          type: string
        method:
          enum:
          - totp
          - recovery_code
          type: string
        mfa_token:
          type: string
      required:
      - code
      - method
      - mfa_token
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    name: API Support
    url: http://www.github.com/EdwinZhanCN/Lumilio-Photos
  description: Media management system API with asset features
  license:
    name: GPLv3.0
    url: https://opensource.org/licenses/GPL-3.0
  title: Lumilio-Photos API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/auth/mfa:
    get:
      description: Get the authenticated user's MFA status, including TOTP enablement
        and remaining recovery codes.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.MFAStatusDTO'
          description: MFA status retrieved successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get MFA status
      tags:
      - auth
...
  /api/v1/auth/mfa/recovery-codes/regenerate:
    post:
      description: Generate a fresh set of recovery codes for the authenticated user.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.RegenerateRecoveryCodesRequestDTO'
                description: Recovery code regeneration payload
                summary: request
        description: Recovery code regeneration payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.RecoveryCodesResponseDTO'
          description: Recovery codes regenerated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: MFA is not enabled
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized or incorrect password
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Regenerate recovery codes
      tags:
      - auth
  /api/v1/auth/mfa/totp/disable:
    post:
      description: Disable TOTP MFA and invalidate recovery codes for the authenticated
        user.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.DisableTOTPRequestDTO'
                description: Disable TOTP payload
                summary: request
        description: Disable TOTP payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.MFAStatusDTO'
          description: TOTP disabled successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: MFA is not enabled
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized or incorrect password
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Disable TOTP
      tags:
      - auth
  /api/v1/auth/mfa/totp/enable:
    post:
      description: Verify a TOTP setup code and enable TOTP MFA for the authenticated
        user.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.EnableTOTPRequestDTO'
                description: TOTP enable payload
                summary: request
        description: TOTP enable payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.RecoveryCodesResponseDTO'
          description: TOTP enabled successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid setup token or verification code
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Enable TOTP
      tags:
      - auth
  /api/v1/auth/mfa/totp/setup:
    post:
      description: Generate a new TOTP secret and setup token for the authenticated
        user.
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.TOTPSetupResponseDTO'
          description: TOTP setup created successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Begin TOTP setup
      tags:
      - auth
  /api/v1/auth/mfa/verify:
    post:
      description: Verify a pending MFA login challenge with a TOTP code or recovery
        code.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.VerifyMFARequestDTO'
                description: MFA verification payload
                summary: request
        description: MFA verification payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.AuthResponseDTO'
          description: MFA verification successful
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid request data
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Invalid or expired MFA challenge
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal server error
      summary: Verify MFA challenge
      tags:
      - auth
...
servers:
- url: localhost:6680/api/v1